+++ src/controllers/originController.ts
@@ -5,6 +5,10 @@ import { Op, col, fn, where } from 'sequelize';
 class OriginController {
   async getAllOrigins(req: Request, res: Response) {
     try {
+      const { tenantId } = req.query;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       const { search } = req.query;
       const filters: any = {};
       if (search) {
@@ -18,7 +22,7 @@ class OriginController {
         ];
       }
 
-      // Llamar al servicio con los filtros
+      filters.tenantId = tenantId;
       const origins = await OriginService.getAllOrigin(filters);
 
       res.status(200).json(origins);
@@ -30,7 +34,11 @@ class OriginController {
 
   async getOriginById(req: Request, res: Response) {
     try {
-      const origin = await OriginService.getOriginById(+req.params.id);
+      const { tenantId } = req.query;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
+      const origin = await OriginService.getOriginById(+req.params.id, +tenantId);
       if (!origin) {
         return res.status(404).json({ message: 'Origin not found' });
       }
@@ -42,6 +50,10 @@ class OriginController {
 
   async createOrigin(req: Request, res: Response) {
     try {
+      const { tenantId } = req.body;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       const newOrigin = await OriginService.createOrigin(req.body);
       res.status(201).json(newOrigin);
     } catch (error: any) {
@@ -51,6 +63,10 @@ class OriginController {
 
   async updateOrigin(req: Request, res: Response) {
     try {
+      const { tenantId } = req.body;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       const updatedOrigin = await OriginService.updateOrigin(+req.params.id, req.body);
       res.status(200).json(updatedOrigin);
     } catch (error: any) {
@@ -60,6 +76,10 @@ class OriginController {
 
   async deleteOrigin(req: Request, res: Response) {
     try {
+      const { tenantId } = req.query;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       await OriginService.deleteOrigin(+req.params.id);
       res.status(204).send();
     } catch (error: any) {
