+++ src/controllers/truckController.ts
@@ -4,8 +4,11 @@ import { Op, literal } from 'sequelize';
 
 class TruckController {
   async getAllTrucks(req: Request, res: Response) {
-    console.log('asdjkasdj')
     try {
+      const { tenantId } = req.query;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
         const { search } = req.query;
 
         const filters: any = {};
@@ -18,8 +21,7 @@ class TruckController {
           ];
         }
 
-        console.log('filters ', filters)
-
+        filters.tenantId = tenantId;
         const trucks = await truckService.getAllTrucks(filters);
         res.status(200).json(trucks);
     } catch (error: any) {
@@ -29,7 +31,11 @@ class TruckController {
 
   async getTruckById(req: Request, res: Response) {
     try {
-      const truck = await truckService.getTruckById(+req.params.id);
+      const { tenantId } = req.query;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
+      const truck = await truckService.getTruckById(+req.params.id, +tenantId);
       if (!truck) {
         return res.status(404).json({ message: 'Truck not found' });
       }
@@ -41,6 +47,10 @@ class TruckController {
 
   async createTruck(req: Request, res: Response) {
     try {
+      const { tenantId } = req.body;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       const newTruck = await truckService.createTruck(req.body);
       res.status(201).json(newTruck);
     } catch (error: any) {
@@ -50,7 +60,10 @@ class TruckController {
 
   async updateTruck(req: Request, res: Response) {
     try {
-      console.log('hahaha asd ', req.body, ' id: ', req.params.id)
+      const { tenantId } = req.body;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       const updatedTruck = await truckService.updateTruck(+req.params.id, req.body);
       res.status(200).json(updatedTruck);
     } catch (error: any) {
@@ -60,6 +73,10 @@ class TruckController {
 
   async deleteTruck(req: Request, res: Response) {
     try {
+      const { tenantId } = req.query;
+      if (!tenantId) {
+          return res.status(400).json({ message: 'Tenant ID is required' });
+      }
       await truckService.deleteTruck(+req.params.id);
       res.status(204).send();
     } catch (error: any) {
